ai_camera_agent/requirements.txt
opencv-python==4.10.0.84
fastapi==0.115.5
numpy==2.2.2
uvicorn==0.32.1
httpx==0.27.2
websockets==12.0
requests~=2.32.5
pydantic~=2.12.5
ultralytics~=8.4.6
python-dotenv
aiosqlite~=0.20.0